Final Updates for Master
