+++ src/edu/ucsb/cs56/projects/games/battleship/view/BattleshipGUI.java
@@ -14,7 +14,9 @@ import javax.sound.sampled.*;
  * @author Wenjian Li (working since W14)
  * @author Chang Rey Tang (W16)
  * @author Joseph Song (W16)
- * @version 2.2 (CS56 Winter 2016)
+ * @author Barbara Korycki (F16)
+ * @author Edward Guardado (F16)
+ * @version 2.3 (CS56 Fall 2016)
 
  */
 
@@ -24,81 +26,44 @@ public class BattleshipGUI extends JFrame{
     private String difficulty = null;
    	private int gameType = 0;
     private int replayType = 0;
-	private int lastMove;
-	private int xLoc;
-	private int yLoc;
-	private int boatToPlace;
+
 	private int boatSpawn;
-    private Color shipColor = Color.DARK_GRAY;
-	private boolean playersTurn = false;
-	private boolean ipEntered = false;
 	private boolean replay = true;
 	private boolean prompt = true;
-	private boolean placeBoats = false;
-	private boolean boatPlaced = false;
-	private boolean horzOrVert = true; //true for horizontal false for verticle
-    private boolean audio = true;
 
     //Public so that other classes can play sound files
-    public URL placeURL;
+
     public URL shotURL;
     public URL missURL;
     public URL winURL;
     public URL loseURL;
-    public URL cantPlaceURL;
-    public Clip clip;
 	
-	//GUI's knowledge bank. Used for GameGrid cell coloring
-	private ArrayList<Integer> playerBoats = new ArrayList<Integer>();
-	private ArrayList<Integer> enemyBoats = new ArrayList<Integer>();
-	private ArrayList<Integer> shots = new ArrayList<Integer>();
-    private ArrayList<ArrayList<Integer>> playerBoatGroups = new ArrayList<ArrayList<Integer>>();
-	public ArrayList<Integer> getEnemyBoats() {return this.enemyBoats;}
+    private JPanel audioPanel = new JPanel();
 
 	//GUI Texts
    	private JLabel title = new JLabel("Battleship",JLabel.CENTER);
 	private JLabel messages = new JLabel("Messages go here:", JLabel.CENTER);
 	
 	//Gametype frame popup
-	private JFrame typePopUp = new JFrame();
-	private JButton hostButton = new JButton("Host a Game");
-	private JButton joinButton = new JButton("Join a Game");
-	private JButton computerButton = new JButton("Play Against a Computer");
+	private TypeSetUpFrame typePopUp = new TypeSetUpFrame();
 	
 	//Difficulty frame popup
-    private JFrame diffPopUp = new JFrame();
-    private JButton easyButton = new JButton("Easy");
-    private JButton mediumButton = new JButton("Medium");
-    private JButton hardButton = new JButton("Hard");
+    private DiffPopUpFrame diffPopUp = new DiffPopUpFrame();
 
 	//Select ship sizes frame popup
-	private JFrame shipSizePopUp = new JFrame();
-	private JButton reminder = new JButton("Input sizes of ships (between 2 and 9), then click to proceed");
-	private JPanel shipSizePanel = new JPanel();
-	private JTextField ship1 = new JTextField(5);
-	private JTextField ship2 = new JTextField(5);
-	private JTextField ship3 = new JTextField(5);
-	private JTextField ship4 = new JTextField(5);
-	private JTextField ship5 = new JTextField(5);
-	private JTextField [] inputBoxes = {ship1, ship2, ship3, ship4, ship5};
+	private SizeSetUpFrame shipSizePopUp = new SizeSetUpFrame();
 
     //Color selection frame popup 
-    private JFrame colorPopUp = new JFrame();
-    private JButton colorSelectButton = new JButton("Select a color, then click to proceed");
-    String [] colorList = {"Default", "Black", "Blue", "Green", "Orange", "Pink", "White", "Yellow"};
-    private JComboBox colorDrop = new JComboBox(colorList);
+    private ColorPopUpFrame colorPopUp = new ColorPopUpFrame();
 	
 	//Join IP frame popup
-	private JFrame ipPopUp = new JFrame();
-	private JLabel ipRequest = new JLabel("Please input the IP address you wish to join.", JLabel.CENTER);
-	private JTextField ipField = new JTextField();
-	private JLabel ipMessage = new JLabel("Hit enter to submit.", JLabel.CENTER);
+	private IpPopUpFrame ipPopUp = new IpPopUpFrame();
 	
     //Play again popup
     private JFrame playAgainPopUp = new JFrame();
     private JFrame networkPlayAgainPopUp = new JFrame();
     private JButton playAgainButton = new JButton("Play Again");
-    private JButton networkPlayAgainButton = new JButton("Play Again");
+    private JButton networkPlayAgainButton = new JButton("Play Again"); //shows up when playing mult
     private JButton newShipsButton = new JButton("Play Again w/ New Ship Sizes");
     private JButton playAgainWithNewIPButton = new JButton( "Join a new game");
     private JButton playAgainAsHostButton = new JButton ("Start a new game as a host");
@@ -107,10 +72,7 @@ public class BattleshipGUI extends JFrame{
 
     //Game board component
     private GameGrid board = new GameGrid();
-
-    //Audio muted/unmuted checkbox
-    private JPanel audioPanel = new JPanel();
-    private JCheckBox audioMute = new JCheckBox("Mute");
+    GridLayout threeButtonGrid = new GridLayout(1,3);
 
     /**
      * Default constructor for the class. Sets everything up.
@@ -123,21 +85,16 @@ public class BattleshipGUI extends JFrame{
         this.getContentPane().add(BorderLayout.NORTH,title);
 
         //Add game board
-        this.board.setSize(100,210);
-        this.board.addMouseListener(this.new cellClick());
-        this.board.addMouseMotionListener(this.new mouseMove());
-        this.board.addKeyListener(this.new changeOrientation());
         this.getContentPane().add(BorderLayout.CENTER,board);
         this.getContentPane().setBackground(Color.WHITE);
 
-        //Add audio controls
+        //Add  controls
         audioPanel.setLayout(new BorderLayout());
         audioPanel.setBackground(Color.WHITE);
         this.getContentPane().add(BorderLayout.EAST,audioPanel);
-        audioMute.setBackground(Color.WHITE);
-        audioPanel.add(audioMute, BorderLayout.SOUTH);
-        audioMute.addItemListener(new audioCheck());
-
+        board.audioMute.setBackground(Color.WHITE);
+        audioPanel.add(board.audioMute, BorderLayout.SOUTH);
+        board.audioMute.addItemListener(board.new audioCheck());
 
         //Add messages
         this.getContentPane().add(BorderLayout.SOUTH, messages);
@@ -146,43 +103,13 @@ public class BattleshipGUI extends JFrame{
         this.setSize(350,500);
         this.board.requestFocusInWindow();
 
-        //setup difficulty options popup
-        GridLayout threeButtonGrid = new GridLayout(1,3);
-        this.diffPopUp.setLayout(threeButtonGrid);
-        this.diffPopUp.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        this.diffPopUp.setSize(600,100);
-
-        //Add difficulty buttons listeners
-        this.easyButton.addActionListener(this.new difficultyClick());
-        this.mediumButton.addActionListener(this.new difficultyClick());
-        this.hardButton.addActionListener(this.new difficultyClick());
-
-        this.diffPopUp.add(easyButton);
-        this.diffPopUp.add(mediumButton);
-        this.diffPopUp.add(hardButton);
-
-        //Setup gametype popup
-        this.typePopUp.setLayout(threeButtonGrid);
-        this.typePopUp.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        this.typePopUp.setSize(800,100);
-
-        //Add type buttons listeners
-        this.hostButton.addActionListener(this.new typeClick());
-        this.joinButton.addActionListener(this.new typeClick());
-        this.computerButton.addActionListener(this.new typeClick());
-
-        //Add type buttons to window
-        this.typePopUp.add(hostButton);
-        this.typePopUp.add(joinButton);
-        this.typePopUp.add(computerButton);
-
         //Setup playAgain popup
         this.playAgainPopUp.setLayout(threeButtonGrid);
         this.playAgainPopUp.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         this.playAgainPopUp.setSize(800,100);
 
-	//Setup networkPlayAgain popup
-	GridLayout fourButtonGrid = new GridLayout(1,4);
+	   //Setup networkPlayAgain popup
+	   GridLayout fourButtonGrid = new GridLayout(1,4);
         this.networkPlayAgainPopUp.setLayout(fourButtonGrid);
         this.networkPlayAgainPopUp.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         this.networkPlayAgainPopUp.setSize(800,100);
@@ -191,9 +118,9 @@ public class BattleshipGUI extends JFrame{
         this.playAgainButton.addActionListener(this.new playAgainClick());
         this.newShipsButton.addActionListener(this.new playAgainClick());
         this.mainMenuButton.addActionListener(this.new playAgainClick());
-	this.playAgainWithNewIPButton.addActionListener(this.new networkPlayAgainClick());
-	this.playAgainAsHostButton.addActionListener(this.new networkPlayAgainClick());
-	this.networkMainMenuButton.addActionListener(this.new networkPlayAgainClick());
+	    this.playAgainWithNewIPButton.addActionListener(this.new networkPlayAgainClick());
+	   this.playAgainAsHostButton.addActionListener(this.new networkPlayAgainClick());
+	   this.networkMainMenuButton.addActionListener(this.new networkPlayAgainClick());
         this.networkPlayAgainButton.addActionListener(this.new networkPlayAgainClick());
 
         //Add playAgain buttons to window
@@ -201,149 +128,161 @@ public class BattleshipGUI extends JFrame{
         this.playAgainPopUp.add(newShipsButton);
         this.playAgainPopUp.add(mainMenuButton);
 
-	 //Add networkPlayAgain buttons to window
-	this.networkPlayAgainPopUp.add(networkPlayAgainButton);
+	    //Add networkPlayAgain buttons to window
+	    this.networkPlayAgainPopUp.add(networkPlayAgainButton);
         this.networkPlayAgainPopUp.add(playAgainWithNewIPButton);
         this.networkPlayAgainPopUp.add(playAgainAsHostButton);
-	this.networkPlayAgainPopUp.add(networkMainMenuButton);
-        
-        //Setup shipsize popup
-        this.shipSizePopUp.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        this.shipSizePopUp.setSize(600,100);
-
-        //Add shipsize buttons listeners
-        this.reminder.addActionListener(this.new sizeClick());
-        for(int i=0; i<5; i++){
-            this.inputBoxes[i].addActionListener(this.new sizeClick());
-        }
-        this.shipSizePopUp.getContentPane().add(BorderLayout.SOUTH, reminder);
-        for(int i=0; i<5; i++){
-            this.shipSizePanel.add(this.inputBoxes[i]);
-        }
-        this.shipSizePopUp.getContentPane().add(BorderLayout.CENTER, shipSizePanel);
-
-        //Setup colorPopUp
-        this.colorPopUp.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        this.colorPopUp.setSize(600,100);
-        
-        //Add shipColor buttons and dropdown listeners
-        this.colorSelectButton.addActionListener(this.new colorClick());
-        this.colorDrop.addActionListener(this.new colorDropSelect());
-
-        //Add buttons and dropdown to window
-        this.colorPopUp.getContentPane().add(BorderLayout.CENTER, colorDrop);
-        this.colorPopUp.getContentPane().add(BorderLayout.SOUTH, colorSelectButton);
-
-        //Setup IP popup
-        GridLayout threeWidgetVerticleGrid = new GridLayout(3,1);
-        this.ipPopUp.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        this.ipPopUp.setLayout(threeWidgetVerticleGrid);
-        this.ipPopUp.setSize(600,100);
-        this.ipField.setHorizontalAlignment(JTextField.CENTER);
-        this.ipField.addActionListener(this.new ipEnter());
-
-        //Add IP widgets
-        this.ipPopUp.getContentPane().add(BorderLayout.NORTH,ipRequest);
-        this.ipPopUp.getContentPane().add(BorderLayout.CENTER,ipField);
-        this.ipPopUp.getContentPane().add(BorderLayout.SOUTH,ipMessage);
-        
+	    this.networkPlayAgainPopUp.add(networkMainMenuButton);
+            
         //Initialize sound file locations
-        placeURL = this.getClass().getResource("sfx/ship_place.aiff");
+        board.placeURL = this.getClass().getResource("sfx/ship_place.aiff");
         shotURL = this.getClass().getResource("sfx/ship_hit.aiff");
         missURL = this.getClass().getResource("sfx/miss_splash.aiff");
         winURL = this.getClass().getResource("sfx/victory.aiff");
         loseURL = this.getClass().getResource("sfx/failure.aiff");
-        cantPlaceURL = this.getClass().getResource("sfx/ship_cant_place.aiff");
+        board.cantPlaceURL = this.getClass().getResource("sfx/ship_cant_place.aiff");
     }
-
-    /**
-     * Main method used only for testing purposes.
-     *@param args not used
-     **/
-        public static void main(String[] args){
-
-            Player human = new Player();
-            human.randomGenerateBoats();
-            BattleshipGUI gui = new BattleshipGUI();
-
-            gui.setOptions();
-            gui.setVisible(true);
-
-            ArrayList<Integer> boats = human.getBoatsArrayList();
-            gui.addPlayerBoats(boats);
-
-            Computer computer = new Computer("HARD");
-            ArrayList<Integer> boatList = computer.getBoatsArrayList();
-            gui.addEnemyBoats(boatList);
-
-            for(int i = 0; i<20; i++){
-                int move = computer.makeMove();
-                gui.addShot(move);
-                computer.updateGuessGrid( move, gui.hitEnemy(move));
-            }
-
-        }
-
-        /**
-         * Returns whether a shot from the player against the enemy is a "HIT" or a "MISS".
-         * @param shot The player's shot.
-         * @return "HIT" or "MISS"
-         **/
-        public String hitEnemy(int shot){
-            if( enemyBoats.contains(shot)) return "HIT";
-            else return "MISS";
-        }
-
-        /**
-         * Returns whether a shot from the enemy against the player is a "HIT" or a "MISS".
-         * @param shot The enemy's shot.
-         * @return "HIT" or "MISS"
-         **/
-
-        public String hitPlayer(int shot){
-            if( playerBoats.contains(shot)) return "HIT";
-            else return "MISS";
-        }
-
-        public void placeBoats() {
-            this.placeBoats = true;
-            int[] boats = Player.shipSizes; 
-            for(int boat: boats){
-                this.boatPlaced = false;
-                this.boatToPlace = boat;
-                while(!this.boatPlaced){
-                    /*try{
-                      Thread.sleep(10);
-                      }
-                      catch (InterruptedException e){
-                      }*/
-                    BattleshipController.sleep();
-                }
-            }
-            this.placeBoats = false;
-            this.repaint();
-        }
-
-        /**
-         * Shifts some location to player's GameGrid by moving it down 11 rows.
-         * @param loc The location to be shifted.
-         * @return shifted integer location value
-         **/
-
-        public int shiftToPlayerGrid(int loc){
-            int boatRow = loc/10 + 11;
-            int boatColumn = loc%10;
-            return boatRow*10 + boatColumn;
-        }
 
         /**
          * Method to initiate option prompts to the user.
+         * This is the start of what makes the GUI go.
          **/
 
         public void setOptions(){
             this.replay = false;
             this.setVisible(false);
             this.typePopUp.setVisible(true);
+            while (!typePopUp.entered){
+                try{
+                    Thread.sleep(100);}
+                catch(InterruptedException e){
+                    System.out.println(e);
+                }
+            }
+            this.gameType = typePopUp.getGameType();
+
+            this.typePopUp.setVisible(false);
+            if(gameType == 1){                 //Host Game
+                this.setVisible(true);
+                this.setIpEntered(true);
+            }
+            else if (gameType == 2){            //Join Game
+                setUpJoinGame();
+            }
+            else if (gameType == 3){            //Play computer
+                setUpComputerGame();
+            }
+        }
+
+        
+        /**
+         * Method to wait on the listeners for the color of the boats.
+         **/
+        public void waitForColor(){
+            while(colorPopUp.getColor() == null || !colorPopUp.getContinue()){
+                try{
+                    Thread.sleep(100);}
+                catch(InterruptedException e){
+                    System.out.println(e);
+                }                    
+                
+            }
+        }
+
+
+        /**
+         * Method to wait for the boat sizes to be given.
+         **/
+        public void waitForSizes(){
+            while(!shipSizePopUp.getSizeSet()){
+                try{
+                    Thread.sleep(100);}
+                catch(InterruptedException e){
+                    System.out.println(e);
+                }                    
+                
+            }
+        }
+
+
+        /**
+         * Method to wait for the difficulty to be chosen.
+         **/
+        public void waitForDifficulty(){
+            while(diffPopUp.getDifficulty() == null){
+                try{
+                    Thread.sleep(100);}
+                catch(InterruptedException e){
+                    System.out.println(e);
+                }                    
+                
+            }            
+        }
+
+
+        /**
+         * Method to set up a computer game and the corresponding graphics.
+         **/
+        public void setUpComputerGame(){
+            this.colorPopUp.setVisible(true);
+            waitForColor();
+            this.board.setShipColor(this.colorPopUp.getColor());
+            this.shipSizePopUp.setVisible(true);           
+            this.colorPopUp.setVisible(false);  
+
+            waitForSizes();
+            board.setShipSizes(shipSizePopUp.getShipSizes());
+            this.shipSizePopUp.setVisible(false);
+            this.diffPopUp.setVisible(true);
+
+            waitForDifficulty();
+            difficulty = diffPopUp.getDifficulty();
+            this.diffPopUp.setVisible(false);
+            this.setVisible(true);
+        }
+
+        
+        /**
+         * Method to initiate the graphics for computer boats.
+         **/
+        public void setUpComputerNewShips(){
+            this.setVisible(false);
+            this.shipSizePopUp.setVisible(true);
+            waitForSizes();
+            board.setShipSizes(shipSizePopUp.getShipSizes());
+            this.shipSizePopUp.setVisible(false);
+            this.diffPopUp.setVisible(true);
+
+            waitForDifficulty();
+            difficulty = diffPopUp.getDifficulty();
+            this.diffPopUp.setVisible(false);
+            this.setVisible(true);
+        }
+
+        /**
+         * Method to initiate the graphics a second game with the AI.
+         **/
+        public void setUpCompterPlayAgain(){
+            waitForDifficulty();
+            difficulty = diffPopUp.getDifficulty();
+            this.diffPopUp.setVisible(false);
+            this.setVisible(true);
+        }
+
+        /**
+         * Method to initiate the graphics for a Joining Game.
+         **/
+        public void setUpJoinGame(){
+            this.ipPopUp.setVisible(true);
+            while(this.ipPopUp.getIpEntered() == false){
+                try{
+                    Thread.sleep(100);}
+                catch(InterruptedException e){
+                    System.out.println(e);
+                }
+            }
+            this.ipPopUp.setVisible(false);
+            this.setVisible(true);
         }
 
         /**
@@ -353,16 +292,10 @@ public class BattleshipGUI extends JFrame{
         public void reset(){
             this.difficulty = null;
             this.gameType = 0;
-            this.playersTurn = false;
-            this.ipEntered = false;
             this.replay = true;
             this.prompt = true;
-            this.lastMove = 0;
 
-            this.playerBoats = new ArrayList<Integer>();
-            this.enemyBoats = new ArrayList<Integer>();
-            this.shots= new ArrayList<Integer>();
-
+            board.reset();
             this.setVisible(false);
         }
 
@@ -370,37 +303,67 @@ public class BattleshipGUI extends JFrame{
         public void resetPlace() {
             this.difficulty = null;
             this.prompt = true;
-            this.lastMove = 0;
             this.gameType = 3;
-            this.playersTurn = false;
-            this.ipEntered = false;
             this.replay = true;
              
-            this.playerBoats = new ArrayList<Integer>();
-            this.enemyBoats = new ArrayList<Integer>();
-            this.shots = new ArrayList<Integer>();
+            board.reset();
         
             this.setVisible(false);
             this.diffPopUp.setVisible(true);
+            setUpCompterPlayAgain();
+        
+        }
+
+        public void resetForIP() {
+            this.prompt = true;
+            this.gameType = 2;
+            this.ipPopUp.setIpEntered(false);
+            this.replay = true;
+             
+            board.reset();
+        
+            this.setVisible(false);
+            setUpJoinGame();
+        
+        }
+
+        public void resetForJoinAgain() {
+            this.setVisible(false);
+            this.prompt = true;
+            this.gameType = 2;
+            this.ipPopUp.setIpEntered(true);
+            this.replay = true;
+             
+            board.reset();
+        
+            this.setVisible(true);
+        
+        }
+        
+        public void resetForHost() {
+            this.setVisible(false);
+            this.prompt = true;
+            this.gameType = 1;
+            this.ipPopUp.setIpEntered(true);
+            this.replay = true;
+             
+            board.reset();
+            setDefaultShipSizes();
+            this.setVisible(true);
         
         }
 
         public void resetShips() {
             this.difficulty = null;
             this.prompt = true;
-            this.lastMove = 0;
             this.gameType = 3; 
-            this.playersTurn = false;
-            this.ipEntered = false;
+            this.ipPopUp.setIpEntered(false);
             this.replay = true;
 
 
-            this.playerBoats = new ArrayList<Integer>();
-            this.enemyBoats = new ArrayList<Integer>();
-            this.shots = new ArrayList<Integer>();
+            board.reset();
 
-            this.setVisible(false);
-            this.shipSizePopUp.setVisible(true);
+            setUpComputerNewShips();
         }
 
         public void computerPlayAgain() {
@@ -408,10 +371,8 @@ public class BattleshipGUI extends JFrame{
         }
 
         public void networkPlayAgain() {
-	    this.networkPlayAgainPopUp.setVisible(true);
-	}
-
-
+	       this.networkPlayAgainPopUp.setVisible(true);
+	   }
 
         /**
          * Changes the title at the top of the gui.
@@ -431,15 +392,6 @@ public class BattleshipGUI extends JFrame{
             this.messages.setText(message);
         }
 
-        /**
-         * Setter for playersTurn variable. Used by controller to let user make a move.
-         * @param tf The boolean value to set to playersTurn.
-         **/
-
-        public void setPlayersTurn(boolean tf){
-            this.playersTurn = tf;
-        }
-	
         /**
          * Getter for replay instance variable. Used to check if the player wants to keep playing.
          * @return true for keep playing false for stop playing
@@ -467,15 +419,6 @@ public class BattleshipGUI extends JFrame{
             return this.difficulty;
         }
 
-        /**
-         * Getter for playersTurn instance variable
-         * @return value stored in playersTurn
-         */
-
-        public boolean getPlayersTurn(){
-            return this.playersTurn;
-        }
-
 	/**
 	 * Returns the message being displayed at the bottom of the GUI.
 	 * @return message String
@@ -491,7 +434,7 @@ public class BattleshipGUI extends JFrame{
 	 **/
 	 
 	public String getIP(){
-		return this.ipField.getText();
+		return this.ipPopUp.getText();
 	}
 	
 	/**
@@ -500,9 +443,24 @@ public class BattleshipGUI extends JFrame{
 	 **/
 	
 	public boolean getIPEntered(){
-		return this.ipEntered;
+		return this.ipPopUp.getIpEntered();
 	}
 	
+    public void setIpEntered(boolean set){
+        this.ipPopUp.setIpEntered(set);
+    }
+    public void setDefaultShipSizes(){
+        int[] array = {2,3,3,4,5};
+        board.setShipSizes(array);
+       // player.setDefaultShipSizes();
+    }
+    public boolean shipSizePopUpVisibile(){
+        return this.shipSizePopUp.isVisible();
+    }
+    public boolean colorPopUpVisible(){
+        return this.colorPopUp.isVisible();
+    }
+
 	/**
 	 * Method for returning status of user prompt
 	 * @return true for the uses acknowledged prompt, false for user hasn't acknowledged prompt
@@ -512,356 +470,26 @@ public class BattleshipGUI extends JFrame{
 		return this.prompt;
 	 }
 	 
-	/**
-	 * Lets gui know its players turn
-	 */
-	
-	public void makeMove(){
-		this.playersTurn = true;
-	}
-	
-	/**
-	 * Check if a spawn is valid for placing boat
-	 * @param spawn representing the specific spawn
-	 */
-
-	public boolean isValidSpawn(int spawn){
-		if(this.horzOrVert){
-			for(int i=0; i < boatToPlace; i++){
-				if((spawn + i)%10 > 9 || playerBoats.contains(spawn+i) || (spawn + i)/10 != spawn/10){
-                    playAudioFile(cantPlaceURL);
-                    return false;
-                    
-                }
-			}
-		}
-		else{
-			for(int i=0; i<boatToPlace; i++){
-				if((spawn+10*i)/10 > 20 || playerBoats.contains(spawn+10*i)) {
-                    playAudioFile(cantPlaceURL);
-                    return false;
-                }
-			}
-		}
-		return true;
-	}
-	
-	public void placeBoat(int spawn){
-	    ArrayList<Integer> spawns = new ArrayList<Integer>();
-		if(horzOrVert){
-		    for(int i=0; i<this.boatToPlace; i++){
-				this.playerBoats.add(spawn + i);
-				spawns.add(spawn + i);
-		    }
-		}
-		else{
-			for(int i=0; i<this.boatToPlace; i++){
-				this.playerBoats.add(spawn + 10*i);
-				spawns.add(spawn + 10*i);
-			}
-		}
-		this.playerBoatGroups.add(spawns);
-	}
-	
-	/**
-	 * Add a shot to the gui's shots list
-	 * @param shot The shot to be added
-	 */
-	
-	public void addShot(int shot){
-		this.shots.add(shot);
-		this.repaint();
-	}
-	
-	/**
-	 * Adds locations for player's boats to the playerBoats list. Shifts their integer locations.
-	 * @param boatList A list of boat locations.
-	 */
-	
-	public void addPlayerBoats(ArrayList<Integer> boatList){
-		for( Integer loc: boatList)
-			this.playerBoats.add(shiftToPlayerGrid(loc));
-	}
-	
-	/**
-	 * Adds locations for enemy's boats to the enemyBoats list.
-	 * @param boatList A list of boat locations.
-	 */
-	
-	public void addEnemyBoats(ArrayList<Integer> boatList){
-		this.enemyBoats = boatList;
-	}
-	
-	/**
-	 * Adds a single boat location to enemyBoats
-	 **/
-	
-	public void addEnemyBoat(int boatLoc){
-		this.enemyBoats.add(boatLoc);
-	}
-	
-	/**
-	 * Method for retrieving player boats. Used when GUI is used to place boats.
-	 **/
-	
-	public ArrayList<Integer> getPlayerBoats(){
-		return this.playerBoats;
-	}
-
-    /**
-     *controller class uses this method to set a player's boat list array
-     **/
-    public ArrayList<ArrayList<Integer>> getGroupBoats(){
-	return this.playerBoatGroups;
-    }
-	
-	/**
-	 * Returns the player's most recent move.
-	 **/
-	
-	public int getLastMove(){
-		return this.lastMove;
-	}
     
     public void end() {
         this.setVisible(false);
         this.getContentPane().removeAll();
     }
 	
-	/**
-	 * Inner class that paints the literal GameGrid for the game.
-	 **/
-	
-    public class GameGrid extends JComponent{
-        public int width;
-        public int height;
-        public int cellWidth;
-        public int startX;
-
-        @Override
-            public void paintComponent(Graphics g)
-            {
-
-                width = this.getWidth() + 50; //50 added to accomodate for mute box
-                height = this.getHeight();		
-                cellWidth = height/21;
-                startX = (width - (cellWidth*10))/2;
-                Graphics2D g2d = (Graphics2D) g;
-                Color ocean = new Color(0,119,190);
-
-                //Make the background white
-                g2d.setColor(Color.WHITE);
-                g2d.fillRect(0,0,this.getWidth(),this.getHeight());
-
-                //Paint individual cells
-                for(int i=0; i < 10; i++){
-                    for(int j = 0; j<21; j++){
-                        int loc = j*10 + i;
-                        if(j==10)
-                            g2d.setColor(Color.BLACK);
-                        else if(shots.contains(loc) && (playerBoats.contains(loc) || enemyBoats.contains(loc))){
-                            g2d.setColor(Color.RED);
-                        }
-                        else if(playerBoats.contains(loc))
-                            g2d.setColor(shipColor);
-                        else if(shots.contains(loc)){
-                            g2d.setColor(ocean.darker());
-                        }
-                        else g2d.setColor(ocean);
-
-                        g2d.fillRect(startX + (i*cellWidth),j*cellWidth,cellWidth,cellWidth);
-
-                    }
-                }
-
-                //Paint GameGrid lines
-                g2d.setColor(Color.GRAY);
-                g2d.drawLine(startX,0,startX,cellWidth*21); //Far left border
-                g2d.drawLine(startX + 10*cellWidth,0,startX + 10*cellWidth,cellWidth*21); //Far right border
-                for(int i=1; i<10; i++)
-                    g2d.drawLine(startX + i*cellWidth,0,startX + cellWidth*i,cellWidth*10);	
-                for(int i=1; i<10; i++)
-                    g2d.drawLine(startX + i*cellWidth,cellWidth*11,startX + cellWidth*i,cellWidth*21);	
-                for(int j=0; j<22; j++)
-                    g2d.drawLine(startX, j*cellWidth, startX + 10*cellWidth, j*cellWidth);
-
-                //Paint boat to be placed if boats are being placed
-                int topLeftX = xLoc - cellWidth/2;
-                int topLeftY = yLoc - cellWidth/2;
-                if(placeBoats && horzOrVert){ //Draw boat to be placed horizontally
-                    g2d.setColor(shipColor);
-                    g2d.fillRect(topLeftX, topLeftY, boatToPlace*cellWidth, cellWidth);
-
-                    g2d.setColor(Color.GRAY);
-                    g2d.drawLine( topLeftX, topLeftY, topLeftX + cellWidth*boatToPlace, topLeftY);
-                    g2d.drawLine( topLeftX, topLeftY + cellWidth, topLeftX + cellWidth*boatToPlace, topLeftY + cellWidth);
-                    for(int i =0; i<boatToPlace + 1; i++){
-                        g2d.drawLine( topLeftX + i*cellWidth, topLeftY, topLeftX + i*cellWidth, topLeftY + cellWidth);
-                    }
-                }
-                else if(placeBoats){ //Draw boat to be places vertically 
-                    g2d.setColor(shipColor);
-                    g2d.fillRect( topLeftX, topLeftY, cellWidth, boatToPlace*cellWidth);
-
-                    g2d.setColor(Color.GRAY);
-                    g2d.drawLine( topLeftX, topLeftY, topLeftX, topLeftY + boatToPlace*cellWidth);
-                    g2d.drawLine( topLeftX + cellWidth, topLeftY, topLeftX + cellWidth, topLeftY + boatToPlace*cellWidth);
-                    for(int i = 0; i< boatToPlace + 1; i++){
-                        g2d.drawLine( topLeftX, topLeftY + i*cellWidth, topLeftX + cellWidth, topLeftY + i*cellWidth);
-                    }
-                }
-            }
-
-
-    }
-
-    /**
-	 * Listener for difficulty option buttons
-	 **/
-	
-    public class difficultyClick implements ActionListener{
-        public void actionPerformed(ActionEvent e){
-            if( e.getSource() == BattleshipGUI.this.easyButton){
-                difficulty = "EASY";
-                BattleshipGUI.this.diffPopUp.setVisible(false);
-                BattleshipGUI.this.setVisible(true);
-            }
-            else if( e.getSource() == BattleshipGUI.this.mediumButton){
-                difficulty = "MEDIUM";
-                BattleshipGUI.this.diffPopUp.setVisible(false);
-                BattleshipGUI.this.setVisible(true);
-            }
-            else if ( e.getSource() == BattleshipGUI.this.hardButton){
-                difficulty = "HARD";
-                BattleshipGUI.this.diffPopUp.setVisible(false);
-                BattleshipGUI.this.setVisible(true);
-            }
-        }
-    }
-	
-	/**
-	 * Listener for the type options buttons
-	 **/
-	 
-    public class typeClick implements ActionListener{
-		public void actionPerformed(ActionEvent e){
-			if( e.getSource() == BattleshipGUI.this.hostButton){
-                gameType = 1;
-                difficulty = "INTERNET";
-                BattleshipGUI.this.typePopUp.setVisible(false);
-                BattleshipGUI.this.setVisible(true);
-            }
-            else if( e.getSource() == BattleshipGUI.this.joinButton){
-                gameType = 2;
-                difficulty = "INTERNET";
-                BattleshipGUI.this.typePopUp.setVisible(false);
-                BattleshipGUI.this.ipPopUp.setVisible(true);
-            }
-            else if ( e.getSource() == BattleshipGUI.this.computerButton){
-                gameType = 3;
-                BattleshipGUI.this.typePopUp.setVisible(false);
-                BattleshipGUI.this.colorPopUp.setVisible(true);
-			}
-		}
-	}
-
-	/**
-	 * Listener for the ship sizes button
-     * once the user clicks the proceed button, texts in the 5 text fields will be check by isValid() method, 
-     * and if all of them are valid, program will read in the 5 numbers into the shipSizes array and 
-     * send the message of clicking to GUI
-	 **/
-
-	public class sizeClick implements ActionListener{
-		public void actionPerformed(ActionEvent e){
-			if(e.getSource() == BattleshipGUI.this.reminder){
-				int [] inputSizes = Player.shipSizes; 
-				boolean isValid = true;
-				
-				for(int i=0; i<5; i++){
-					String input = BattleshipGUI.this.inputBoxes[i].getText();
-					if(BattleshipGUI.isValid(input)==true) {inputSizes[i] = Integer.parseInt(BattleshipGUI.this.inputBoxes[i].getText());}
-					else {isValid=false; break;}
-				}
-				if(isValid==true){
-					BattleshipGUI.this.shipSizePopUp.setVisible(false);
-					BattleshipGUI.this.diffPopUp.setVisible(true);
-				}
-				else reminder.setText("Please input sizes between 2 and 9");
-			}
-		}
-	}
-
-    /**
-    * Listener for the mute check box
-    * audio is muted when checked and unmuted when unchecked
-    **/
-
-    public class audioCheck implements ItemListener{
-        public void itemStateChanged(ItemEvent e){
-            JCheckBox cb = (JCheckBox) e.getSource();
-            if(cb.isSelected())
-                audio = false;
-            else
-                audio = true;
-        }
-    }
-
-    /**
-     * Listener for the color selection menu's continue button
-     * when the user clicks the button it will move onto the ship size
-     * selection menu
-     **/
-
-    public class colorClick implements ActionListener{ 
-        public void actionPerformed(ActionEvent e){
-            if (e.getSource() == BattleshipGUI.this.colorSelectButton){
-                BattleshipGUI.this.colorPopUp.setVisible(false);
-                BattleshipGUI.this.shipSizePopUp.setVisible(true);
-            }
-        }
-    }
-
-    /**
-    * Listener for the color selection menu's dropdown button
-    * when the user selects a color it will set shipColor to that color
-    **/
-
-    public class colorDropSelect implements ActionListener{
-        public void actionPerformed(ActionEvent e){
-            JComboBox combo = (JComboBox)e.getSource();
-            switch((String)combo.getSelectedItem()){
-                case "Default":
-                    shipColor = Color.DARK_GRAY;
-                    break;
-                case "Black":
-                    shipColor = Color.BLACK;
-                    break;
-                case "Blue":
-                    shipColor = Color.BLUE;
-                    break;
-                case "Green":
-                    shipColor = Color.GREEN;
-                    break;
-                case "Orange":
-                    shipColor = Color.ORANGE;
-                    break;
-                case "Pink":
-                    shipColor = Color.PINK;
-                    break;
-                case "White":
-                    shipColor = Color.WHITE;
-                    break;
-                case "Yellow":
-                    shipColor = Color.YELLOW;
-                    break;
-                default:
-                    break;
-            }
-        }
-    }
-    
-
+	public void placeBoats(){ board.placeBoats(); }
+    public ArrayList<Integer> getPlayerBoats(){ return board.getPlayerBoats();}
+    public void addPlayerBoats(ArrayList<Integer> boatList){ board.addPlayerBoats(boatList);}
+    public void addEnemyBoats(ArrayList<Integer> boatList){board.addEnemyBoats(boatList);}
+    public void makeMove(){board.makeMove();}
+    public boolean getPlayersTurn(){ return board.getPlayersTurn();}
+    public int getLastMove(){ return board.getLastMove();}
+    public ArrayList<Integer> getEnemyBoats() {return board.getEnemyBoats();}
+    public void playAudioFile(URL audioURL){    board.playAudioFile(audioURL);}
+    public int shiftToPlayerGrid(int loc){return board.shiftToPlayerGrid(loc);}
+    public void addShot(int shot){ board.addShot(shot);}
+    public String hitPlayer(int shot){ return board.hitPlayer(shot);}
+    public void addEnemyBoat(int boatLoc){ board.addEnemyBoat(boatLoc);}
+     
     /**
      * Listener for the play again options
      **/
@@ -904,7 +532,7 @@ public class BattleshipGUI extends JFrame{
                 BattleshipGUI.this.networkPlayAgainPopUp.setVisible(false);
             }
 
-	    else if(e.getSource() == BattleshipGUI.this.playAgainAsHostButton) {
+	       else if(e.getSource() == BattleshipGUI.this.playAgainAsHostButton) {
                 replayType = 6;
                 BattleshipGUI.this.setVisible(false);
                 BattleshipGUI.this.networkPlayAgainPopUp.setVisible(false);
@@ -917,99 +545,4 @@ public class BattleshipGUI extends JFrame{
             }
         } 
     }    
-
-	/**
-	 * Method exlusively for ship sizes button listener to check validity of user input
-	 **/
-	public static boolean isValid(String input){
-		if(input.equals("") || input.length()!=1) return false;
-		char character = input.charAt(0);
-		if(character>'9' || character<'2') return false;
-		return true;
-	}
-	
-	/**
-	 * Listener class for entering IP addresses
-	 **/
-	
-	public class ipEnter implements ActionListener{
-		public void actionPerformed(ActionEvent e){
-			ipEntered = true;
-			BattleshipGUI.this.ipPopUp.setVisible(false);
-			BattleshipGUI.this.setVisible(true);
-		}
-	}
-	
-	/**
-	 * Mouse listener for clicking on game cells
-	 **/
-    public class cellClick implements MouseListener{
-		public void mouseClicked(MouseEvent e){}
-		public void mousePressed(MouseEvent e){
-			//Calculate which column & row the mouse was clicked in
-			int cellColumn = (int) Math.floor((e.getX() - board.startX)/board.cellWidth);
-			int cellRow = (int) Math.floor(e.getY()/board.cellWidth);
-			//Add to shot list if it was in enemy territory
-			if(cellRow < 10 && cellRow >=0 && cellColumn >=0 && cellColumn < 10 && playersTurn && (!shots.contains(cellRow*10 + cellColumn))){
-				lastMove = cellRow*10 + cellColumn;
-				shots.add(lastMove);
-				playersTurn = false;
-			}
-			//Record click for boat placement if it was in player's territory
-			else if( cellRow < 21 && cellRow > 10 && cellColumn >=0 && cellColumn < 10){
-				if(placeBoats){
-					int spawn = cellRow*10 + cellColumn;
-					if(isValidSpawn(spawn)){ //Place the boat on click if the spawn location is valid
-						boatPlaced = true;
-						placeBoat(spawn);
-    
-                        if(audio)
-                            playAudioFile(placeURL);
-    
-						BattleshipGUI.this.repaint();
-					}
-				}
-			}
-			repaint();
-		}
-		public void mouseReleased(MouseEvent e){}
-		public void mouseEntered(MouseEvent e){}
-		public void mouseExited(MouseEvent e){BattleshipGUI.this.repaint();}
-    }
-	
-	public class mouseMove implements MouseMotionListener{
-		public void mouseMoved(MouseEvent e){
-			if(placeBoats){ //Records x & y locations if player is placing boats
-				xLoc = e.getX();
-				yLoc = e.getY();
-				BattleshipGUI.this.repaint();
-			}
-		}
-		public void mouseDragged(MouseEvent e){}
-	}
-
-	public class changeOrientation implements KeyListener{
-		public void keyPressed(KeyEvent e){
-			horzOrVert = !horzOrVert; //Switch from drawing boats horizontally to drawing them vertically
-			BattleshipGUI.this.repaint();
-		}	
-		public void keyReleased(KeyEvent e){}
-		public void keyTyped(KeyEvent e){}
-	}
-
-    /**  
-    * Methods that plays audio clip referenced by audioURL
-    **/
-    public void playAudioFile(URL audioURL){
-        try{
-            AudioInputStream audioStream = AudioSystem.getAudioInputStream(audioURL);
-            this.clip = AudioSystem.getClip();
-            this.clip.open(audioStream);
-            clip.start();
-            clip.setMicrosecondPosition(0);
-        } catch(Exception e) {
-            System.err.println(e);
-        } 
-    }
-
 }
